Parse the GA urls returned as nextLink on response
When GA returns filters that use an AND (represented as a semicolon) it
does not escape it as %3B. This does not agree with urlparse, as it is
by the book. In order to not strip AND filters we should make sure all
the semicolons are escaped before parsing the url.